+++ src/main/java/com/codeborne/selenide/Selenide.java
@@ -2,49 +2,29 @@ package com.codeborne.selenide;
 
 import com.codeborne.selenide.ex.DialogTextMismatch;
 import com.codeborne.selenide.ex.JavaScriptErrorsFound;
-import com.codeborne.selenide.impl.BySelectorCollection;
 import com.codeborne.selenide.impl.DownloadFileWithHttpRequest;
 import com.codeborne.selenide.impl.ElementFinder;
+import com.codeborne.selenide.impl.JavascriptErrorsCollector;
+import com.codeborne.selenide.impl.Modals;
 import com.codeborne.selenide.impl.Navigator;
-import com.codeborne.selenide.impl.SelenideFieldDecorator;
-import com.codeborne.selenide.impl.WebElementsCollectionWrapper;
-
-import org.openqa.selenium.Alert;
+import com.codeborne.selenide.impl.SelenideWait;
+import com.codeborne.selenide.impl.WebDriverLogs;
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebDriverException;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.logging.LogEntry;
-import org.openqa.selenium.support.PageFactory;
-import org.openqa.selenium.support.ui.FluentWait;
 
 import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Constructor;
 import java.net.URL;
-import java.time.Duration;
-import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.logging.Level;
-import java.util.logging.Logger;
 
-import static com.codeborne.selenide.Configuration.captureJavascriptErrors;
-import static com.codeborne.selenide.Configuration.dismissModalDialogs;
-import static com.codeborne.selenide.Configuration.pollingInterval;
 import static com.codeborne.selenide.Configuration.timeout;
 import static com.codeborne.selenide.WebDriverRunner.closeWebDriver;
 import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
-import static com.codeborne.selenide.WebDriverRunner.hasWebDriverStarted;
-import static com.codeborne.selenide.WebDriverRunner.supportsJavascript;
-import static com.codeborne.selenide.WebDriverRunner.supportsModalDialogs;
 import static com.codeborne.selenide.impl.WebElementWrapper.wrap;
-import static java.util.Arrays.asList;
-import static java.util.Collections.emptyList;
 
 /**
  * The main starting point of Selenide.
@@ -53,10 +33,12 @@ import static java.util.Collections.emptyList;
  * {@link #$(String)} for searching web elements.
  */
 public class Selenide {
-  private static final Logger log = Logger.getLogger(Selenide.class.getName());
-
   public static Navigator navigator = new Navigator();
-
+  private static JavascriptErrorsCollector javascriptErrorsCollector = new JavascriptErrorsCollector();
+  private static WebDriverLogs webDriverLogs = new WebDriverLogs();
+  private static Modals modals = new Modals();
+  private static SelenidePageFactory pageFactory = new SelenidePageFactory();
+  private static DownloadFileWithHttpRequest downloadFileWithHttpRequest = new DownloadFileWithHttpRequest();
 
   /**
    * The main starting point in your tests.
@@ -96,7 +78,7 @@ public class Selenide {
    */
   public static void open(String relativeOrAbsoluteUrl, String domain, String login, String password) {
     navigator.open(relativeOrAbsoluteUrl, domain, login, password);
-    mockModalDialogs();
+    modals.mockModalDialogs();
   }
 
   /**
@@ -138,7 +120,7 @@ public class Selenide {
    */
   public static void open(URL absoluteUrl, String domain, String login, String password) {
     navigator.open(absoluteUrl, domain, login, password);
-    mockModalDialogs();
+    modals.mockModalDialogs();
   }
 
   /**
@@ -150,27 +132,6 @@ public class Selenide {
   public static void updateHash(String hash) {
     String localHash = (hash.charAt(0) == '#') ? hash.substring(1) : hash;
     executeJavaScript("window.location.hash='" + localHash + "'");
-  }
-
-  private static boolean doDismissModalDialogs() {
-    return !supportsModalDialogs() || dismissModalDialogs;
-  }
-
-  private static void mockModalDialogs() {
-    if (doDismissModalDialogs()) {
-      String jsCode =
-          "  window._selenide_modalDialogReturnValue = true;\n" +
-          "  window.alert = function(message) {};\n" +
-          "  window.confirm = function(message) {\n" +
-          "    return window._selenide_modalDialogReturnValue;\n" +
-          "  };";
-      try {
-        executeJavaScript(jsCode);
-      }
-      catch (UnsupportedOperationException cannotExecuteJsAgainstPlainTextPage) {
-        log.warning(cannotExecuteJsAgainstPlainTextPage.toString());
-      }
-    }
   }
 
   /**
@@ -331,7 +292,7 @@ public class Selenide {
    */
   @Deprecated
   public static SelenideElement $(WebElement parent, String cssSelector) {
-    return ElementFinder.wrap($(parent), By.cssSelector(cssSelector), 0);
+    return ElementFinder.wrap(parent, cssSelector);
   }
 
   /**
@@ -342,7 +303,7 @@ public class Selenide {
    * @return SelenideElement
    */
   public static SelenideElement $(String cssSelector, int index) {
-    return ElementFinder.wrap(null, By.cssSelector(cssSelector), index);
+    return ElementFinder.wrap(cssSelector, index);
   }
 
   /**
@@ -359,7 +320,7 @@ public class Selenide {
    */
   @Deprecated
   public static SelenideElement $(WebElement parent, String cssSelector, int index) {
-    return ElementFinder.wrap($(parent), By.cssSelector(cssSelector), index);
+    return ElementFinder.wrap(parent, cssSelector, index);
   }
 
   /**
@@ -397,11 +358,9 @@ public class Selenide {
 
   /**
    * Initialize collection with Elements
-   * @param elements
-   * @return
    */
   public static ElementsCollection $$(Collection<? extends WebElement> elements) {
-    return new ElementsCollection(new WebElementsCollectionWrapper(elements));
+    return new ElementsCollection(elements);
   }
 
   /**
@@ -415,7 +374,7 @@ public class Selenide {
    * @return empty list if element was no found
    */
   public static ElementsCollection $$(String cssSelector) {
-    return new ElementsCollection(new BySelectorCollection(By.cssSelector(cssSelector)));
+    return new ElementsCollection(cssSelector);
   }
 
   /**
@@ -428,7 +387,7 @@ public class Selenide {
    * @return ElementsCollection which locates elements via XPath
    */
   public static ElementsCollection $$x(String xpathExpression) {
-    return new ElementsCollection(new BySelectorCollection(By.xpath(xpathExpression)));
+    return $$(By.xpath(xpathExpression));
   }
 
   /**
@@ -442,7 +401,7 @@ public class Selenide {
    * @return empty list if element was no found
    */
   public static ElementsCollection $$(By seleniumSelector) {
-    return new ElementsCollection(new BySelectorCollection(seleniumSelector));
+    return new ElementsCollection(seleniumSelector);
   }
 
   /**
@@ -462,7 +421,7 @@ public class Selenide {
    */
   @Deprecated
   public static ElementsCollection $$(WebElement parent, String cssSelector) {
-    return new ElementsCollection(new BySelectorCollection(parent, By.cssSelector(cssSelector)));
+    return new ElementsCollection(parent, cssSelector);
   }
 
   /**
@@ -476,7 +435,7 @@ public class Selenide {
    */
   @Deprecated
   public static ElementsCollection $$(WebElement parent, By seleniumSelector) {
-    return new ElementsCollection(new BySelectorCollection(parent, seleniumSelector));
+    return new ElementsCollection(parent, seleniumSelector);
   }
 
   /**
@@ -533,8 +492,7 @@ public class Selenide {
 
   /**
    * Returns selected element in radio group
-   * @param radioField
-   * @return null, if nothing selected
+   * @return null if nothing selected
    */
   public static SelenideElement getSelectedRadio(By radioField) {
     for (WebElement radio : $$(radioField)) {
@@ -550,9 +508,7 @@ public class Selenide {
    * @param confirmReturnValue true = OK, false = CANCEL
    */
   public static void onConfirmReturn(boolean confirmReturnValue) {
-    if (doDismissModalDialogs()) {
-      executeJavaScript("window._selenide_modalDialogReturnValue = " + confirmReturnValue + ';');
-    }
+    modals.onConfirmReturn(confirmReturnValue);
   }
 
   /**
@@ -560,7 +516,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String confirm() {
-    return confirm(null);
+    return modals.confirm();
   }
 
   /**
@@ -571,14 +527,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String confirm(String expectedDialogText) {
-    if (!doDismissModalDialogs()) {
-      Alert alert = switchTo().alert();
-      String actualDialogText = alert.getText();
-      alert.accept();
-      checkDialogText(expectedDialogText, actualDialogText);
-      return actualDialogText;
-    }
-    return null;
+    return modals.confirm(expectedDialogText);
   }
 
   /**
@@ -586,7 +535,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String prompt() {
-    return prompt(null, null);
+    return modals.prompt();
   }
 
   /**
@@ -595,7 +544,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String prompt(String inputText) {
-    return prompt(null, inputText);
+    return modals.prompt(inputText);
   }
 
   /**
@@ -607,16 +556,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String prompt(String expectedDialogText, String inputText) {
-    if (!doDismissModalDialogs()) {
-      Alert alert = switchTo().alert();
-      String actualDialogText = alert.getText();
-      if (inputText != null)
-        alert.sendKeys(inputText);
-      alert.accept();
-      checkDialogText(expectedDialogText, actualDialogText);
-      return actualDialogText;
-    }
-    return null;
+    return modals.prompt(expectedDialogText, inputText);
   }
 
 
@@ -625,7 +565,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String dismiss() {
-    return dismiss(null);
+    return modals.dismiss();
   }
 
   /**
@@ -636,21 +576,7 @@ public class Selenide {
    * @return actual dialog text
    */
   public static String dismiss(String expectedDialogText) {
-    if (!doDismissModalDialogs()) {
-      Alert alert = switchTo().alert();
-      String actualDialogText = alert.getText();
-      alert.dismiss();
-      checkDialogText(expectedDialogText, actualDialogText);
-      return actualDialogText;
-    }
-    return null;
-  }
-
-  private static void checkDialogText(String expectedDialogText, String actualDialogText) {
-    if (expectedDialogText != null && !expectedDialogText.equals(actualDialogText)) {
-      Screenshots.takeScreenShot(Selenide.class.getName(), Thread.currentThread().getName());
-      throw new DialogTextMismatch(actualDialogText, expectedDialogText);
-    }
+    return modals.dismiss(expectedDialogText);
   }
 
   /**
@@ -675,26 +601,17 @@ public class Selenide {
   }
 
   /**
-   * Create a Page Object instance.
-   * @see PageFactory#initElements(WebDriver, Class)
+   * Create a Page Object instance
    */
   public static <PageObjectClass> PageObjectClass page(Class<PageObjectClass> pageObjectClass) {
-    try {
-      Constructor<PageObjectClass> constructor = pageObjectClass.getDeclaredConstructor();
-      constructor.setAccessible(true);
-      return page(constructor.newInstance());
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to create new instance of " + pageObjectClass, e);
-    }
+    return pageFactory.page(getWebDriver(), pageObjectClass);
   }
 
   /**
-   * Create a Page Object instance.
-   * @see PageFactory#initElements(WebDriver, Class)
+   * Initialize a given Page Object instance
    */
   public static <PageObjectClass, T extends PageObjectClass> PageObjectClass page(T pageObject) {
-    SelenidePageFactory.initElements(new SelenideFieldDecorator(getWebDriver()), pageObject);
-    return pageObject;
+    return pageFactory.page(getWebDriver(), pageObject);
   }
 
   /**
@@ -707,10 +624,8 @@ public class Selenide {
    *
    * @return instance of org.openqa.selenium.support.ui.FluentWait
    */
-  public static FluentWait<WebDriver> Wait() {
-    return new FluentWait<>(getWebDriver())
-      .withTimeout(Duration.of(timeout, ChronoUnit.MILLIS))
-      .pollingEvery(Duration.of(pollingInterval, ChronoUnit.MILLIS));
+  public static SelenideWait Wait() {
+    return new SelenideWait(getWebDriver());
   }
 
   /**
@@ -742,60 +657,11 @@ public class Selenide {
    * @return list of error messages. Returns empty list if webdriver is not started properly.
    */
   public static List<String> getJavascriptErrors() {
-    if (!captureJavascriptErrors) {
-      return emptyList();
-    }
-    else if (!hasWebDriverStarted()) {
-      return emptyList();
-    }
-    else if (!supportsJavascript()) {
-      return emptyList();
-    }
-    try {
-      Object errors = executeJavaScript("return window._selenide_jsErrors");
-      if (errors == null) {
-        return emptyList();
-      }
-      else if (errors instanceof List) {
-        return errorsFromList((List<Object>) errors);
-      }
-      else if (errors instanceof Map) {
-        return errorsFromMap((Map<Object, Object>) errors);
-      }
-      else {
-        return asList(errors.toString());
-      }
-    } catch (WebDriverException | UnsupportedOperationException cannotExecuteJs) {
-      log.warning(cannotExecuteJs.toString());
-      return emptyList();
-    }
-  }
-
-  private static List<String> errorsFromList(List<Object> errors) {
-    if (errors.isEmpty()) {
-      return emptyList();
-    }
-    List<String> result = new ArrayList<>(errors.size());
-    for (Object error : errors) {
-      result.add(error.toString());
-    }
-    return result;
-  }
-
-  private static List<String> errorsFromMap(Map<Object, Object> errors) {
-    if (errors.isEmpty()) {
-      return emptyList();
-    }
-    List<String> result = new ArrayList<>(errors.size());
-    for (Map.Entry error : errors.entrySet()) {
-      result.add(error.getKey() + ": " + error.getValue());
-    }
-    return result;
+    return javascriptErrorsCollector.getJavascriptErrors();
   }
 
   /**
    * Check if there is not JS errors on the page
-   * @throws JavaScriptErrorsFound
    */
   public static void assertNoJavascriptErrors() throws JavaScriptErrorsFound {
     List<String> jsErrors = getJavascriptErrors();
@@ -853,7 +719,7 @@ public class Selenide {
    * @see java.util.logging.Level
    */
   public static List<String> getWebDriverLogs(String logType, Level logLevel) {
-    return listToString(getLogEntries(logType, logLevel));
+    return webDriverLogs.getWebDriverLogs(getWebDriver(), logType, logLevel);
   }
 
   /**
@@ -884,22 +750,13 @@ public class Selenide {
     return executeJavaScript("return navigator.userAgent;");
   }
 
-  private static List<LogEntry> getLogEntries(String logType, Level logLevel) {
-    try {
-      return getWebDriver().manage().logs().get(logType).filter(logLevel);
-    }
-    catch (UnsupportedOperationException ignore) {
-      return emptyList();
-    }
-  }
-
   /**
    * Return true if bottom of the page is reached
    *
    * Useful if you need to scroll down by x pixels unknown number of times.
    */
   public static boolean atBottom() {
-    return Selenide.executeJavaScript("return window.pageYOffset + window.innerHeight >= document.body.scrollHeight");
+    return executeJavaScript("return window.pageYOffset + window.innerHeight >= document.body.scrollHeight");
   }
 
   /**
@@ -916,17 +773,6 @@ public class Selenide {
   }
 
   public static File download(String url, long timeoutMs) throws IOException {
-    return new DownloadFileWithHttpRequest().download(url, timeoutMs);
-  }
-
-  private static <T> List<String> listToString(List<T> objects) {
-    if (objects == null || objects.isEmpty()) {
-      return emptyList();
-    }
-    List<String> result = new ArrayList<>(objects.size());
-    for (T object : objects) {
-      result.add(object.toString());
-    }
-    return result;
+    return downloadFileWithHttpRequest.download(url, timeoutMs);
   }
 }