+++ src/test/java/com/codeborne/selenide/impl/NavigatorTest.java
@@ -70,43 +70,10 @@ class NavigatorTest implements WithAssertions {
   }
 
   @Test
-  void addsRandomNumbersToEveryUrlToAvoidIECaching() {
-    assertThat(navigator.makeUniqueUrlToAvoidIECaching("http://chuck-norris.com", 666))
-      .isEqualTo("http://chuck-norris.com?timestamp=666");
-
-    assertThat(navigator.makeUniqueUrlToAvoidIECaching("http://chuck-norris.com?timestamp=123456789", 666))
-      .isEqualTo("http://chuck-norris.com?timestamp=666");
-
-    assertThat(navigator.makeUniqueUrlToAvoidIECaching("http://chuck-norris.com?timestamp=123456789#", 666))
-      .isEqualTo("http://chuck-norris.com?timestamp=666");
-
-    assertThat(navigator.makeUniqueUrlToAvoidIECaching("http://chuck-norris.com?timestamp=123456789&abc=def", 666))
-      .isEqualTo("http://chuck-norris.com?timestamp=666");
-  }
-
-  @Test
   void returnsAbsoluteUrl() {
     Configuration.baseUrl = "http://localhost:8080";
     assertThat(navigator.absoluteUrl("/users/id=1")).isEqualTo("http://localhost:8080/users/id=1");
     assertThat(navigator.absoluteUrl("http://host:port/users/id=1")).isEqualTo("http://host:port/users/id=1");
-  }
-
-  @Test
-  void appendBasicAuthToURL_absoluteUrl() {
-    assertThat(navigator.appendBasicAuthToURL("https://company.com/login", "domain-01", "login-01", "password-01"))
-      .isEqualTo("https://domain-01%5Clogin-01:password-01@company.com/login");
-  }
-
-  @Test
-  void appendBasicAuthToURL_relativeUrl() {
-    assertThat(navigator.appendBasicAuthToURL("/login", "domain-01", "login-01", "password-01"))
-      .isEqualTo("domain-01%5Clogin-01:password-01@/login");
-  }
-
-  @Test
-  void appendBasicAuthToURL_emptyDomain() {
-    assertThat(navigator.appendBasicAuthToURL("https://company.com/login", "", "login-01", "password-01"))
-      .isEqualTo("https://login-01:password-01@company.com/login");
   }
 
   @Test