+++ src/test/java/integration/errormessages/MethodCalledOnCollectionPassesOnTest.java
@@ -1,13 +1,12 @@
 package integration.errormessages;
 
-import java.util.Arrays;
-
 import com.codeborne.selenide.Configuration;
 import com.codeborne.selenide.ElementsCollection;
 import integration.IntegrationTest;
-import integration.helpers.HTMLBuilderForTestPreconditions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
 
 import static com.codeborne.selenide.CollectionCondition.exactTexts;
 import static com.codeborne.selenide.Condition.cssClass;
@@ -17,7 +16,7 @@ import static com.codeborne.selenide.Selenide.$$;
 class MethodCalledOnCollectionPassesOnTest extends IntegrationTest {
   @BeforeEach
   void openPage() {
-    HTMLBuilderForTestPreconditions.Given.openedPageWithBody(
+    givenHtml(
       "<ul>Hello to:",
       "<li class='the-expanse detective'>Miller</li>",
       "<li class='the-expanse missing'>Julie Mao</li>",