+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 4.14.0 (planned to xx.09.2018)
-* Fix #784 Enable BasicAuth through Selenide proxy server  -- see https://github.com/codeborne/selenide/pull/785
+* #784 Enable BasicAuth through Selenide proxy server  -- see https://github.com/codeborne/selenide/pull/785
+* #789 Remove `?timestamp` parameter for IE
 
 ## 4.13.0 (released 20.08.2018)
 * #771 Added method `$.lastChild()` for retrieving the last child element of a given element